cropping/requirements.txt
numpy
pillow
openslide-python